feat: plugin versioning - WIP
