+++ src/hooks/contracts/plugins/useFlipperContract.ts
@@ -3,7 +3,7 @@ import { PluginInfo } from '@/types';
 import { useContract } from 'useink';
 
 export default function useFlipperContract(plugin: PluginInfo) {
-  const metadata = FlipperMetadatas.find((m) => m.hash === plugin.codeHash);
+  const metadata = FlipperMetadatas.find((m) => m.hash === plugin.codeHash || m.version === plugin.version);
 
   return {
     contract: metadata && useContract(plugin.address, metadata.metadata, plugin.chainId),
