+++ src/hooks/contracts/plugins/usePollsContract.ts
@@ -3,7 +3,7 @@ import { PluginInfo } from '@/types';
 import { useContract } from 'useink';
 
 export default function usePollsContract(plugin: PluginInfo) {
-  const metadata = PollsMetadatas.find((m) => m.hash === plugin.codeHash);
+  const metadata = PollsMetadatas.find((m) => m.hash === plugin.codeHash || m.version === plugin.version);
 
   return {
     contract: metadata && useContract(plugin.address, metadata.metadata, plugin.chainId),
