+++ src/hooks/contracts/plugins/usePostsContract.ts
@@ -3,7 +3,7 @@ import { PluginInfo } from '@/types';
 import { useContract } from 'useink';
 
 export default function usePostsContract(plugin: PluginInfo) {
-  const metadata = PostsMetadatas.find((m) => m.hash === plugin.codeHash);
+  const metadata = PostsMetadatas.find((m) => m.hash === plugin.codeHash || m.version === plugin.version);
 
   return {
     contract: metadata && useContract(plugin.address, metadata.metadata, plugin.chainId),
