+++ src/pages/plugins/Posts/0.1.x/index.tsx
@@ -1,0 +1,10 @@
+import { PluginProps } from '@/types';
+
+export default function Posts_V0_1_X({ plugin }: PluginProps) {
+  console.log('Posts_V0_1_X', plugin);
+  return (
+    <div>
+      <h1>Posts_V0_1_x</h1>
+    </div>
+  );
+}
