+++ src/pages/plugins/Posts/0.2.x/PendingPosts/PendingPostsView.tsx
@@ -1,9 +1,9 @@
 import { Box, Text } from '@chakra-ui/react';
-import MembersPendingPostsView from '@/pages/plugins/Posts/PendingPosts/MembersPendingPostsView';
-import SpaceOwnerPendingPostsView from '@/pages/plugins/Posts/PendingPosts/SpaceOwnerPendingPostsView';
-import { usePostsContext } from '@/pages/plugins/Posts/PostsProvider';
 import { useSpaceContext } from '@/providers/SpaceProvider';
 import { MemberStatus, PostPerm } from '@/types';
+import { usePostsContext } from '../PostsProvider';
+import MembersPendingPostsView from './MembersPendingPostsView';
+import SpaceOwnerPendingPostsView from './SpaceOwnerPendingPostsView';
 
 export default function PendingPostsView() {
   const { isOwner, memberStatus } = useSpaceContext();
