+++ src/pages/plugins/Posts/0.2.x/PendingPosts/UpdatePendingPostButton.tsx
@@ -19,48 +19,50 @@ import {
   Tabs,
   TabList,
   Tab,
-  IconButton,
   Link,
+  MenuItem,
 } from '@chakra-ui/react';
 import { FormEvent, useEffect } from 'react';
 import { toast } from 'react-toastify';
-import { useEffectOnce } from 'react-use';
 import { useTx } from '@/hooks/useink/useTx';
-import { usePostsContext } from '@/pages/plugins/Posts/PostsProvider';
 import { Props } from '@/types';
-import { eventEmitter, EventName } from '@/utils/eventemitter';
 import { renderMd } from '@/utils/mdrenderer';
 import { messages } from '@/utils/messages';
 import { notifyTxStatus } from '@/utils/notifications';
-import { AddIcon } from '@chakra-ui/icons';
+import { EditIcon } from '@chakra-ui/icons';
 import { useFormik } from 'formik';
 import { shouldDisableStrict } from 'useink/utils';
-import * as yup from 'yup';
+import { usePostsContext } from '../PostsProvider';
+import { postValidationScheme } from '../actions/NewPostButton';
 
-export const postValidationScheme = yup.object().shape({
-  content: yup.string().required().max(500, 'Content must be at most 500 characters'),
-});
-
-interface NewPostButtonProps extends Props {
-  onPostCreated: () => void;
+interface UpdatePostButtonProps extends Props {
+  postId: number;
+  defaultValue: string;
+  // onPostUpdated: (content: any, postId: number) => void;
+  disabled?: boolean;
 }
 
-export default function NewPostButton({ onPostCreated }: NewPostButtonProps) {
-  const { contract, shouldCreatePendingPost } = usePostsContext();
-  const newPostTx = useTx<number>(contract, 'newPost');
+export default function UpdatePendingPostButton({
+  postId,
+  defaultValue,
+  // onPostUpdated,
+  disabled = false,
+}: UpdatePostButtonProps) {
+  const { contract } = usePostsContext();
+  const updatePostTx = useTx<number>(contract, 'updatePendingPost');
   const { isOpen, onOpen, onClose } = useDisclosure();
 
   const formik = useFormik({
     initialValues: {
-      content: '',
+      content: defaultValue,
     },
     validationSchema: postValidationScheme,
     onSubmit: (values, formikHelpers) => {
       const { content } = values;
       const postContent = { Raw: content };
-      newPostTx.signAndSend([postContent], undefined, (result) => {
+      updatePostTx.signAndSend([postId, postContent], {}, (result) => {
         if (!result) {
-          newPostTx.resetState(formikHelpers);
+          updatePostTx.resetState(formikHelpers);
           return;
         }
 
@@ -70,16 +72,13 @@ export default function NewPostButton({ onPostCreated }: NewPostButtonProps) {
           if (result.dispatchError) {
             toast.error(messages.txError);
           } else {
-            toast.success(
-              shouldCreatePendingPost
-                ? 'Your post will be show up after being reviewed by space owner'
-                : 'New post created',
-            );
+            toast.success('Post updated');
 
-            onPostCreated();
+            // Set current content by updated content
+            // onPostUpdated(postContent, postId);
           }
 
-          newPostTx.resetState(formikHelpers);
+          updatePostTx.resetState(formikHelpers);
           onClose();
         }
       });
@@ -87,44 +86,21 @@ export default function NewPostButton({ onPostCreated }: NewPostButtonProps) {
   });
 
   useEffect(() => {
-    newPostTx.resetState();
+    updatePostTx.resetState();
     formik.resetForm();
   }, [isOpen]);
 
-  useEffectOnce(() => {
-    const showPopup = () => onOpen();
-    eventEmitter.on(EventName.SHOW_NEW_POST_POPUP, showPopup);
-
-    return () => {
-      eventEmitter.off(EventName.SHOW_NEW_POST_POPUP, showPopup);
-    };
-  });
-
-  const processing = shouldDisableStrict(newPostTx);
+  const processing = shouldDisableStrict(updatePostTx);
 
   return (
     <>
-      <Button
-        variant='outline'
-        colorScheme='primary'
-        size='sm'
-        onClick={onOpen}
-        display={{ base: 'none', md: 'block' }}>
-        New
-      </Button>
-      <IconButton
-        aria-label={'New post'}
-        colorScheme='primary'
-        variant='outline'
-        size='sm'
-        onClick={onOpen}
-        icon={<AddIcon />}
-        display={{ base: 'block', md: 'none' }}
-      />
+      <MenuItem icon={<EditIcon />} isDisabled={disabled} onClick={onOpen}>
+        Edit
+      </MenuItem>
       <Modal isOpen={isOpen} onClose={onClose} size={{ base: 'full', md: '3xl' }} closeOnOverlayClick={false}>
         <ModalOverlay />
         <ModalContent as='form' onSubmit={(e) => formik.handleSubmit(e as FormEvent<HTMLFormElement>)}>
-          <ModalHeader>New Post</ModalHeader>
+          <ModalHeader>Edit Post #{postId}</ModalHeader>
           <ModalCloseButton />
           <ModalBody>
             <Tabs variant='enclosed' borderStyle='solid' borderWidth={1} borderRadius={4} size='sm'>
@@ -179,9 +155,9 @@ export default function NewPostButton({ onPostCreated }: NewPostButtonProps) {
                 type='submit'
                 width={100}
                 isLoading={processing}
-                loadingText='Posting...'
+                loadingText='Saving...'
                 isDisabled={processing || !formik.isValid}>
-                Post
+                Save
               </Button>
             </Flex>
           </ModalFooter>
