+++ src/pages/plugins/Posts/0.2.x/PostCard.tsx
@@ -2,15 +2,15 @@ import { Box, Flex, IconButton, Menu, MenuButton, MenuList, Text } from '@chakra
 import { Identicon } from '@polkadot/react-identicon';
 import { RiMore2Fill } from 'react-icons/ri';
 import useContractState from '@/hooks/useContractState';
-import PinPostButton from '@/pages/plugins/Posts/actions/PinPostButton';
-import UnpinPostButton from '@/pages/plugins/Posts/actions/UnpinPostButton';
-import UpdatePostButton from '@/pages/plugins/Posts/actions/UpdatePostButton';
 import { useSpaceContext } from '@/providers/SpaceProvider';
 import { useWalletContext } from '@/providers/WalletProvider';
 import { MemberInfo, MemberStatus, PostContent, PostRecord, Props } from '@/types';
 import { fromNow } from '@/utils/date';
 import { renderMd } from '@/utils/mdrenderer';
 import { shortenAddress } from '@/utils/string';
+import PinPostButton from './actions/PinPostButton';
+import UnpinPostButton from './actions/UnpinPostButton';
+import UpdatePostButton from './actions/UpdatePostButton';
 
 interface PostCardProps extends Props {
   postRecord: PostRecord;
